Update cash-register.css
**HTML:**
Removed the "Coins used:" display from the summary.
Removed the <br> from the summary details too.

**CSS:**
Added a max height to the body to keep it from bleeding over the edge and requiring scrolling.
Added the "empty cash" class. This fades important elements when the cash drawer is empty and the till is therefore closed.
Created the button-blocker to stop use of the till when it is closed (Occam's Razor here - rather than rewire the logic, just block the usage!).
Added z-index to the summary display as it was hiding behind the bottom row container now that the latter has a position value.