Update cash-register.js
**HTML:**
Removed the "Coins used:" display from the summary.
Removed the <br> from the summary details too.

**CSS:**
Added a max height to the body to keep it from bleeding over the edge and requiring scrolling.
Added the "empty cash" class. This fades important elements when the cash drawer is empty and the till is therefore closed.
Created the button-blocker to stop use of the till when it is closed (Occam's Razor here - rather than rewire the logic, just block the usage!).
Added z-index to the summary display as it was hiding behind the bottom row container now that the latter has a position value.

**JS:**
Invoked renderStatus and openTill at the renderCashInDrawer function to reset the till status after using set-cash button.
Created function renderCoin to fade the empty (€0) cash drawers or bring them back to normal when they're full again.
Removed list-item elements after closing the summary display (to create a new basket/receipt for a new customer).
Added the summary details to the right-hand summary box (change, cash paid, status, coins used etc).
Fixed it to only open the summary display when the sale goes through (not after insufficient payment/funds or basket empty).
Changed updateCashInDrawer to run not only on ==0 but <=0 just in case it overshoots at some point in the future.
Created two functions to create the li elements needed for the summary details and then to render them there. I also use this function for the sortChange array, to make each coin and the number of said coin to stand on their individual lines.